AppAlgorithmFace/requirements-gh.txt
aliyun-python-sdk-core==2.15.0
aliyun-python-sdk-kms==2.16.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cmake==3.28.3
crcmod==1.7
cryptography==42.0.5
decorator==4.4.2
dlib==19.24.2
ffmpeg-python==0.2.0
filelock==3.13.1
future==1.0.0
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
jdconfig==1.0
Jinja2==3.1.3
jmespath==0.10.0
json5==0.9.24
lit==18.1.1
MarkupSafe==2.1.5
moviepy==1.0.3
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.9.0.80
oss2==2.18.4
pillow==10.2.0
proglog==0.1.10
pycparser==2.21
pycryptodome==3.20.0
requests==2.31.0
six==1.16.0
sympy==1.12
toml==0.10.2
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
tqdm==4.66.2
triton==2.0.0
typing_extensions==4.10.0
urllib3==2.2.1
